python-service/requirements.txt
fastapi==0.116.1
httpx==0.28.1
pydantic==2.11.3
pydantic-settings==2.10.1
PyJWT==2.10.1
uvicorn==0.35.0
pytest==8.4.2
pytest-asyncio==1.1.0
